requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.8.3
django-bootstrap-pagination==1.6.1
django-irs-filings==0.1.6
django-queryset-csv==0.3.1
django-toolbelt==0.0.1
django-watson==1.2.2
future==0.15.0
futures==3.0.3
gunicorn==19.3.0
Metafone==0.5
probablepeople==0.3.1
psycopg2==2.6.2
python-crfsuite==0.8.3
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
six==1.9.0
static3==0.7.0
tomorrow==0.2.3
tzlocal==1.2
